requirements.txt
pandas
tqdm
PyGithub
SQLAlchemy
psycopg2-binary
